Added lefthook and fixed build error
