+++ backend/main.py
@@ -10,8 +10,6 @@ from app.api.controllers.property import router as property_router
 from app.api.controllers.neighbourhood import router as neighbourhood_router
 from app.api.controllers.camera import router as camera_router
 from app.api.controllers.users import router as users_router
-# from app.api.controllers.alerts import router as alerts_router
-from app.api.controllers.alert import router as alert_router
 from app.api.controllers.stream import router as stream_router
 from slowapi.middleware import SlowAPIMiddleware
 from app.auth.rate_limiter import limiter
